Added support for command synonyms
Command synonyms allow commands to operate under multiple keys without having to use seperate scripts or wrappers to point to functions. This is useful for cases where a command may change to a new key but the need to retain the old key is present

commands cannot share keys, if there is a duplicate key the first command to load will recieve that key and the rest will be denied that key.